requirements.txt
pytest==3.4.1
tensorflow_gpu==1.6.0
scipy==1.0.0
pyyaml==3.12
